DRY contact association methods
